refactor: move nunjucks filters into `src/template-filters` directory
