+++ src/template-filters/omitPackageScope.ts
@@ -1,0 +1,15 @@
+import { errorMsgTag } from '../utils/nunjucks';
+
+export function omitPackageScopeName(packageName: string): string;
+export function omitPackageScopeName(packageName: undefined): undefined;
+export function omitPackageScopeName(packageName: string | undefined): string | undefined;
+export function omitPackageScopeName(packageName: string | undefined): string | undefined {
+    return packageName?.replace(/^@[^/]+\//, '');
+}
+
+export function omitPackageScope(packageName: unknown): string {
+    if (typeof packageName !== 'string') {
+        throw new TypeError(errorMsgTag`Invalid packageName value: ${packageName}`);
+    }
+    return omitPackageScopeName(packageName);
+}
